+++ changes.md
@@ -5,6 +5,7 @@
 2. Support TBranchObject - appears in old files
 3. Fix - problem with zooming in THStack
 4. Fix - empty baskets at the end of branch store
+5. Fix - more checks when reading version for empty STL container 
 
 
 ## Changes in 5.0.0